Avatar (#7)
* attempting to receive a file inorder to upload an avatar, early stages

* added a post endpoint in routes, need postman or frontend to pass the file itself in the request body

* trying to successfully receive from frontend to upload file

* multipart not supported potverde

* file upload hell yeaaaa, cleaned up, need to test clean version and add get file endpoint for frontend

* src -> server, same as frontend

* cleaning up um + db avatar controllers + will add get avatar file too

* sorted avatar uploading so request is only parsed in db, everything nice and clean and worky worky

---------
