+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "build": "npx tsc",
     "start": "node dist/server.js",
-    "dev": "nodemon --watch src --ext ts --exec \"tsc --noEmit && npx prisma migrate dev --name initial --schema=./prisma/schema.prisma  && tsx src/server.ts || exit 1\""
+    "dev": "nodemon --watch server --ext ts --exec \"tsc --noEmit && npx prisma migrate dev --name initial --schema=./prisma/schema.prisma  && tsx server/server.ts || exit 1\""
   },
   "keywords": [],
   "author": "",
@@ -25,13 +25,17 @@
   "dependencies": {
     "@fastify/cors": "^11.0.0",
     "@fastify/formbody": "^8.0.2",
+    "@fastify/multipart": "^9.0.3",
     "@fastify/static": "^8.1.1",
     "@prisma/client": "^6.5.0",
     "better-sqlite3": "^11.9.0",
     "close-with-grace": "^2.2.0",
-    "fastify": "^5.2.1"
+    "fastify": "^5.2.1",
+    "fastify-file-upload": "^4.0.0",
+    "form-data": "^4.0.2",
+    "node-fetch": "^3.3.2"
   },
   "prisma": {
-	"schema": "prisma/schema.prisma"
+    "schema": "prisma/schema.prisma"
   }
 }
