+++ prisma/schema.prisma
@@ -28,6 +28,10 @@ model User {
 
   player Player?
 }
+model UploadedFile {
+  id        Int    @id @default(autoincrement())
+  filename  String
+}
 
 model Player {
     id            Int   @id @default(autoincrement())
