+++ tsconfig.json
@@ -6,7 +6,7 @@
 	  "moduleResolution": "NodeNext",
 	  "noEmitOnError": true, // Stops compiling if it runs into errors
 	  "outDir": "./dist",
-	  "rootDir": "./src",
+	  "rootDir": "server",
 	  "strict": true,
 	  "esModuleInterop": true,
 	  "allowSyntheticDefaultImports": true,
@@ -15,6 +15,6 @@
 	  "skipLibCheck": true,
 	  "forceConsistentCasingInFileNames": true
 	},
-	"include": ["src/**/*.ts", "src/server.ts", "src/types.d.ts", "anotes/old_db/addMemberDev.controller.ts", "anotes/old_db/createManyDev.controller.ts", "anotes/old_db/deleteMemberDev.controller.ts"],
+	"include": ["server/**/*.ts", "server/server.ts", "server/types.d.ts", "anotes/old_db/addMemberDev.controller.ts", "anotes/old_db/createManyDev.controller.ts", "anotes/old_db/deleteMemberDev.controller.ts"],
 	"exclude": ["node_modules"]
 }
